athena/requirements.txt
# Google API dependencies
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.0

# Database dependencies
psycopg2-binary>=2.9.0

# Utility dependencies
requests>=2.25.0
